Updated fancyhdr configuration, fixing issues
The previous fancyhdr configuration was a mish-mash of styles that
I applied over the course of time. I've revisited the configuration
this time and made it much easier to understand, as well as moved the
actual package loading into the optional configuration.tex file. This
makes it easy to remove the fancy styles, when desired.