Remove protrusion.lua; issue is fixed upstream
My template had a rather complicated protrusion.lua file which uses the
lua scripting layer of LuaLaTeX to manually adjust the microprotrusion
settings of quotation marks from csquotes, that are within an itemize
or enumerate environment. This issue has since been fixed upstream in
the microtype package as of version 3.0c, so it is not needed anymore.