Added reproducible builds for PDFs using makefile
I ensured that PDFs created using the makefile in this project are
completely bit-for-bit reproducible. In order to accomplish this, I
modified the makefile to use SOURCE_EPOCH_DATE, based on the latest
git commit. All builds have the same timestamp, when they are based
off the same commit, hence PDFs are fully reproducible.