Migrate template to custom latex documentclass
This is the first major update to latex-essay in over a year. In this
commit, we completely remove the template/ directory, and instead move
all the configuration that was previously done in template/packages.tex
and template/formatting.tex into a single unified humanize.cls file.

This new humanize article class, is a custom latex documentclass that is
based off the article documentclass. By migrating our template to a
custom documentclass, we improve the UX and ergonomics of using the
template tremondously. It allows us to expose configuration of the
template as documentclass options, and hide more of the underlying
plumbing from the end user.

With this migration, we also add some additional improvements, such as:
more robust handling of hanging punctuation: we no longer get microtype
errors. We also implement protruded ieee-style bibliography numbers, in
the same style as our protruded section numbers. Finally, we also add
the babel package for multilingual support.

Right now, the documentation has not yet caught up to the template. We
will update the documentation in subsequent commits.