Update documentation to reflect new humanize.cls
This is a quick update which removes the outdated parts of README.md,
and adds a little additional info about using the new humanize.cls file.
I still have to perform a proper writeup of the documentation later,
this just serves as an interim update which will let me share the
repository.