Change package load order so base class opts pass
Here we change the \RequiredPackage load order and move the vast
majority of the packages after the option declarations, so that when we
declare the final, wildcard options handler \DeclareOption*, the
additional options are passed on to the base documentclass. This is
because we must first have \DeclareOption*, and then run \LoadClass,
doing it the other way will not work. Because now that we are loading
our packages after option declaration, we must also wrap a lot of the
option code with an \AtEndOfClass environment so they won't be executed
while they are undefined.