Make quote environment indent same as \parindent
Previously, the indentation (i.e. left and right margins) on the quote
environment were similar, but not quite the same as the \parindent
value. This left a small but noticeable space which looks quite ugly
when an indented paragraph is followed by a quote. By re-defining the
quote environment (which is used by csquotes for \blockquote,
\textquote, etc.), we make the overall look of the blockquote more
uniform.