Updated itemize,enumerate to use \parindent
In a similar change to the commit updating the \quote environment, this
commit modifies the \leftmargin value of itemize, enumerate, and
description environments to use the following calculated value:

    leftmargin = \parindent - \itemindent

This value effectively aligns the text-block of items in itemize and
enumerate environments to be on the same \parindent indent as paragraph
indentation and quotations, leading to a more consistent visual look.

In order to showcase this change, small modifications were also made to
the default example template.